requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy[asyncio]==2.0.30
alembic==1.13.2
aiogram==3.10.0
APScheduler==3.10.4
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
google-auth==2.29.0
google-api-python-client==2.129.0
asyncpg==0.29.0
pytest==8.3.3
pytest-asyncio==0.23.8
google-auth-oauthlib==1.2.1
pytest==8.3.3
pytest-asyncio==0.23.8

# Speech-to-text
vosk==0.3.45
ffmpeg-python==0.2.0

# Mistral AI
mistralai==0.4.1


